cof/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-crispy-forms==2.0
django-extensions==3.2.3
django-import-export==3.2.0
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
MarkupPy==1.14
numpy==1.25.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
Pillow==10.0.0
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
xlrd==2.0.1
xlwt==1.3.0
